requirements.txt
alabaster==0.7.12
anyio==3.6.1
appdirs==1.4.4
async-timeout==4.0.2
Babel==2.10.3
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
esbonio==0.14.1
fastapi==0.85.0
h11==0.14.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pendulum==2.1.2
pptree==3.1
pydantic==1.9.2
pygls==0.12.2
Pygments==2.13.0
pylev==1.4.0
pyparsing==3.0.9
pyspellchecker==0.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-ulid==1.1.0
pytz==2022.4
pytzdata==2020.1
redis==4.3.4
redis-om==0.0.27
requests==2.28.1
rq==1.11.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.20.4
typeguard==2.13.3
types-redis==4.3.21
types-six==1.16.21
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wrapt==1.14.1
